refactor: Update custom type fields for isFeatured in tours and accommodations
The custom type fields for "isFeatured" have been added to the "tours_single" and "accomodation_single" custom types. This allows the content authors to mark certain tours and accommodations as featured. The "isFeatured" field is of type Boolean and has a default value of false. This update provides flexibility in showcasing featured content on the website.