refactor: Add "Show all" link to AccommodationList and TourList slices
The AccommodationList and TourList slices have been updated to include a "Show all" link when the limit is not set to "No limit". This allows users to easily navigate to a page displaying all accommodations or tours. The link is styled with a hover effect and uses the LinkBuilderWithLocale component for localization. This enhancement improves the user experience and provides a convenient way to access all available content.